Simplifies Makefile.

+++ Makefile
@@ -1,22 +1,9 @@
 PWD = `pwd`
-LINKS =".bash_aliases"  \
-	".bash_completion_maven" \
-	".bash_completion_git" \
-	".bash_exports" \
-	".bash_colours" \
-	".bash_extra" \
-	".bash_functions" \
-	".bash_logout" \
-	".bash_path" \
-	".bash_profile" \
-	".bash_prompt" \
-	".bashrc" \
-	".inputrc" \
-	".powerline-bash.py"
+LINKS=$(shell ls -A | grep ^\\.)
 
 all: link
 
 link:
 	for file in $(LINKS); do \
-		ln -sf "$$PWD/$$file" ~/"$$file"; \
+		[ -f $$file ] && [ $$file != ".git" ] && ln -sf "$$PWD/$$file" ~/"$$file"; \
 	done
